# openapi.yaml
openapi: 3.0.0
info:
  title: Indicator Backend API
  version: 1.0.0
  description: |
    API documentation for the Indicator application backend.

    This API provides endpoints for managing educational institutions' data including:
    - User authentication and management
    - School basic information (alapadatok)
    - Student enrollment data (tanulo_letszam)
    - Educational data (tanugyi_adatok)
    - Competency measurements (kompetencia)
    - Admission statistics (felvettek_szama)
    - Table and data access management

    ## Authentication

    Most endpoints require JWT Bearer token authentication. To authenticate:
    1. Use the `/auth/login` endpoint to obtain a token
    2. Include the token in the Authorization header: `Bearer <token>`

    ## Base URL

    All API endpoints are prefixed with `/api/v1`

  contact:
    name: API Support
    url: https://indicator.pollak.info
    email: support@pollak.info
  license:
    name: Private
    url: https://indicator.pollak.info/license
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Invalid input data"
        message: "The provided data does not meet the required format"
...
    Elhelyezkedes:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        szakirany_id:
          type: integer
          nullable: true
          description: Field of study identifier
          example: 1
        szakma_id:
          type: integer
          nullable: true
          description: Profession identifier
          example: 1
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        vegzettek_szama:
          type: integer
          description: Number of graduates
          example: 85
        elhelyezkedett_hallgatok_szama:
          type: integer
          description: Number of placed graduates
          example: 72
        elhelyezkedesek_aranya:
          type: number
          format: float
          description: Placement rate percentage
          example: 84.7
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Last update timestamp
    ElhelyezkedesInput:
      type: object
      required:
        - alapadatok_id
        - tanev_kezdete
        - vegzettek_szama
        - elhelyezkedett_hallgatok_szama
      properties:
        szakirany_id:
          type: integer
          nullable: true
          description: Field of study identifier
          example: 1
        szakma_id:
          type: integer
          nullable: true
          description: Profession identifier
          example: 1
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        vegzettek_szama:
          type: integer
          description: Number of graduates
          example: 85
        elhelyezkedett_hallgatok_szama:
          type: integer
          description: Number of placed graduates
          example: 72

    # HHesHHHNeveluTanulok schemas
...
security:
  - bearerAuth: []
...
paths:
...
  /elhelyezkedes/{tanev}:
    get:
      tags: [Elhelyezkedes]
      summary: Get graduate placement data by school year
      description: Retrieve graduate placement data for a specific school year
      parameters:
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Placement data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Elhelyezkedes"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elhelyezkedes/{alapadatokId}/{tanev}:
    get:
      tags: [Elhelyezkedes]
      summary: Get placement data by school and year
      description: Retrieve graduate placement data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Placement data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Elhelyezkedes"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Elhelyezkedes]
      summary: Delete placement data by school and year
      description: Delete all placement data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "204":
          description: Placement data deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elhelyezkedes:
    post:
      tags: [Elhelyezkedes]
      summary: Create new placement data
      description: Create new graduate placement data entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElhelyezkedesInput"
      responses:
        "201":
          description: Placement data created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elhelyezkedes"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elhelyezkedes/{id}:
    put:
      tags: [Elhelyezkedes]
      summary: Update placement data
      description: Update existing graduate placement data entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Placement record ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElhelyezkedesInput"
      responses:
        "200":
          description: Placement data updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elhelyezkedes"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # HH_es_HHH_nevelesu_tanulok endpoints
...